test: consolidate redundant usage-service tests and split monolith by topic
- Fold the OAuth-403 body variants, resume-trigger pairs, Retry-After
  backoff pair, and does-not-cache-fallback resolver pair into
  parameterized case loops (95 tests -> 88, no pinned behavior lost)
- Drop a resolver test fully subsumed by the interactive-shell
  fallback test
- Split the 3,344-line ClaudeUsageServiceTests.swift into topical
  files: polling/backoff, token recovery, and headers fallback live
  in extensions of a single XCTestCase, with shared spies and
  factories in the base file
- Keep one test class on purpose: Xcode parallelizes per class
  across runner processes and these tests share UserDefaults-backed
  AppSettings state; separate classes raced and failed
- Extract resolver tests into standalone ClaudeCLIResolverTests
  (per-process static hooks only, safe to parallelize) and add it to
  the focused preset in scripts/test.sh
